chore(Types): add *INT() macros for truncating integers into `int` without undefined behaviors
Also rewrite expr to lessen the truncated expr part [R]

* include/cppdef.h:
    * add INT() [I]
    * add COUNTOF_INT() for truncated COUNTOF() [C]
    * add STRLITLEN_INT() for truncated STRLITLEN() [S]
* include/global.h:
    * build_langver [I]
* innbbsd/rec_article.c:
    * receive_article() [I]
* lib/logger.c:
    * loggerf() [I]
* maple/menu.c:
    * menu_foot() [RI]
* maple/more.c:
    * outs_footer() [I]
* maple/talk.c:
    * ulist_body() [I]
    * ulist_neck() [I]
* maple/visio.c:
    * Macro o_ansi() (#ifndef M3_USE_PFTERM) [S]
* maple/xchatd.c:
    * chat_list_rooms() [I]
    * command_execute() [I]
* maple/xover.c:
    * Every_Z_Xover() [I]
* so/adminutil.c:
    * m_resetsys() [I]
* so/chat.c:
    * chat_topic() [RI]
* so/pip.c:
    * pip_do_menu() [S]
    * pip_mainmenu() [S]
    * pip_buy_goods_new() [I]
    * pip_practice_class*() [C]
    * pip_practice_gradeup() [C]
    * pip_go_palace() [S]
    * pip_vs_man() [S]
    * pip_fight_menu() [S]

+++ maple/talk.c
@@ -3189,7 +3189,7 @@ ulist_body(
                 strcpy(color, wcolor[fcolor]);
 
                 /* IID.2021-03-06: Limit the length of the field for user name to make the field for user address information larger */
-                const int margin = BMAX((d_cols >> 1) + 21 - (int)(unsigned)BMAX(sizeof(up->realname), sizeof(up->username)), 0);
+                const int margin = BMAX((d_cols >> 1) + 21 - INT(BMAX(sizeof(up->realname), sizeof(up->username))), 0);
                 const int name_len = (d_cols >> 1) + 21 - margin;
                 const int from_len = ((d_cols+1) >> 1) + 15 + margin;
 
@@ -3413,7 +3413,7 @@ ulist_neck(
     XO *xo)
 {
     /* IID.2021-03-06: Limit the length of the field for user name to make the field for user address information larger */
-    const int margin = BMAX((d_cols >> 1) + 21 - (int)(unsigned)BMAX(MEMBER_SIZE(ACCT, realname), MEMBER_SIZE(ACCT, username)), 0);
+    const int margin = BMAX((d_cols >> 1) + 21 - INT(BMAX(MEMBER_SIZE(ACCT, realname), MEMBER_SIZE(ACCT, username))), 0);
     const int name_len = (d_cols >> 1) + 21 - margin;
     const int from_len = ((d_cols+1) >> 1) + 15 + margin;
